Switch to doubles, use correct alpha in jacobi

+++ src/sycl/SyclSolver.h
@@ -6,9 +6,9 @@ public:
 	static void solve(SyclGridData& grid);
 
 private:
-	static float vcycle(cl::sycl::queue& queue, SyclGridData& grid);
+	static double vcycle(cl::sycl::queue& queue, SyclGridData& grid);
 	static void jacobi(cl::sycl::handler& cgh, SyclGridData& grid, std::size_t levelNum, std::size_t maxiter);
 	static void compResidual(cl::sycl::handler& cgh, SyclGridData& grid, std::size_t levelNum);
-	static float sumResidual(cl::sycl::queue& queue, SyclGridData& grid, std::size_t levelNum);
+	static double sumResidual(cl::sycl::queue& queue, SyclGridData& grid, std::size_t levelNum);
 	static void restrict(cl::sycl::handler& cgh, SyclBuffer& fine, SyclBuffer& coarse);
 };
